backend/requirements.txt
aiosqlite>=0.20,<1
anthropic>=0.39,<1
fastapi>=0.115,<1
google-generativeai>=0.8,<1
httpx>=0.27,<1
mediapipe>=0.10,<0.11
numpy>=1.26,<2
onnxruntime>=1.19,<2
opencv-python>=4.10,<5
Pillow>=10,<12
pydantic>=2.8,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9,<1
requests>=2.32,<3
sentence-transformers>=2.2,<4
spotipy>=2.24,<3
uvicorn[standard]>=0.34,<1
